refactor(entity): Move POJO entity classes to new directory
This will keep the POJO entities separate from the Hibernate entities.
